python_consumer/requirements.txt
# Python Kafka Consumer Dependencies

confluent-kafka==2.3.0
pyyaml==6.0.1

